+++ lint-code.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 # Convenience script for linting code before committing.
 
-isort -rc . && black --skip-string-normalization --line-length 90 .
+isort . && black --skip-string-normalization --line-length 90 .